cache_service/requirements.txt
Werkzeug>=3.0.0  # Compatible con Flask 3
flask==3.0.2
redis==5.0.1
psycopg2-binary==2.9.9